requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.32
botocore==1.34.32
Brotli==1.1.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.4
crispy-bootstrap5==2023.10
cron-descriptor==1.4.3
cssselect2==0.7.0
Django==5.0.3
django-appconf==1.0.6
django-bootstrap-datepicker-plus==5.0.5
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-extensions==3.2.3
django-js-asset==2.2.0
django-model-utils==4.3.1
django-redis==5.4.0
django-sass-processor==1.4
django-scheduler==0.10.1
django-storages==1.14.2
django-timezone-field==6.1.0
django_compressor==4.5
Faker==25.0.1
flower==2.0.1
fonttools==4.50.0
geographiclib==2.0
geopy==2.4.1
gunicorn==21.2.0
html5lib==1.1
humanize==4.9.0
icalendar==5.0.11
idna==3.6
jmespath==1.0.1
kombu==5.3.6
libsass==0.23.0
packaging==23.2
pillow==10.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
pydotplus==2.0.2
pydyf==0.9.0
pygraphviz==1.13
pyparsing==3.1.2
pyphen==0.14.0
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
rcssmin==1.1.2
redis==5.0.1
reportlab==4.1.0
requests==2.31.0
rjsmin==1.2.2
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
stripe==8.1.0
swapper==1.3.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
whitenoise==6.6.0
zopfli==0.2.3
